add: tratamento de exceção ao excluir Os ou Vendas

+++ application/controllers/Os.php
@@ -547,7 +547,11 @@ class Os extends MY_Controller
             redirect(base_url() . 'index.php/os/gerenciar/');
         }
 
-        $this->os_model->delete('cobrancas', 'os_id', $id);
+        if($os->idCobranca != null) {
+            $this->session->set_flashdata('error', 'Existe uma cobrança associada a esta OS, deve cancelar ou excluir a cobrança primeiro!');
+            redirect(site_url('os/gerenciar/'));
+        }
+
         $this->os_model->delete('servicos_os', 'os_id', $id);
         $this->os_model->delete('produtos_os', 'os_id', $id);
         $this->os_model->delete('anexos', 'os_id', $id);
